Change the PN Abstraction to Protocol
