+++ sdk/communication/AzureCommunicationChat/Source/ChatClient.swift
@@ -42,7 +42,7 @@ public class ChatClient {
     private var realTimeNotificationDisconnectedHandler: TrouterEventHandler?
     private var pushNotificationClient: PushNotificationClient?
     internal var registrationId: String
-    public weak var chatClientPushNotificationDelegate: ChatClientPushNotificationDelegate?
+    public weak var keyManager: ChatClientPushNotificationProtocol?
 
     // MARK: Initializers
 
@@ -404,13 +404,12 @@ public class ChatClient {
 
         let encryptionKey: String
 
-        // Delegate "storeKey" behaviour if the Contoso intends to implement encryption
-        if chatClientPushNotificationDelegate != nil {
+        // Persist the key if the Contoso intends to implement encryption
+        if keyManager != nil {
             encryptionKey = generateEncryptionKey()
-            chatClientPushNotificationDelegate?.chatClientPushNotification(
-                _: self,
-                willPersist: encryptionKey,
-                of: Date(timeIntervalSinceNow: 45 * 60)
+            keyManager?.onPersistKey(
+                encryptionKey,
+                expiryTime: Date(timeIntervalSinceNow: 45 * 60)
             )
         } else {
             encryptionKey = ""
@@ -467,63 +466,6 @@ public class ChatClient {
                     .error("Failed to stop push notifications with error: \(error.localizedDescription)")
                 completionHandler(.failure(AzureError.client("Failed to stop push notifications", error)))
             }
-        }
-    }
-
-    /// Handle the data payload for an incoming push notification.
-    /// - Parameters:
-    ///   - notification: The APNS push notification payload ( including "aps" and "data" )
-    ///   - pushNotificationEncryptionDelegate: Manage encyrption keys used by push notification.
-    public func decryptPayload(
-        notification: [AnyHashable: Any]
-    ) throws -> PushNotificationEvent {
-        // Retrieve the "data" part from the APNS push notification payload
-        guard let dataPayload = notification["data"] as? [String: AnyObject] else {
-            throw AzureError.client("Push notification does not contain data payload")
-        }
-
-        do {
-            // 1.get "eventId"
-            guard let eventId = dataPayload["eventId"] as? Int else {
-                throw AzureError
-                    .client("Push notification does not contain eventId or eventId can't be downcast to Int.")
-            }
-
-            let chatEventType = try PushNotificationChatEventType(forCode: eventId)
-
-            // 2.get "e"
-            guard let encryptedPayload = dataPayload["e"] as? String else {
-                throw AzureError
-                    .client(
-                        "Push notification does not contain encryptedPayload or payload can't be downcast to String."
-                    )
-            }
-
-            // 3.Delegate "getKeys" behaviour
-            guard let encryptionKeys = chatClientPushNotificationDelegate?
-                .chatClientPushNotificationWillGetKeys(_: self),
-                encryptionKeys.count > 0
-            else {
-                throw AzureError.client("Failed to get decryption keys. Failed to decrypt Notification payload.")
-            }
-
-            // 4.Verify and decrypt the encrypted notification payload
-            let decryptedPayload = try PushNotificationClient.decryptPayload(
-                encryptedStr: encryptedPayload,
-                encryptionKeys: encryptionKeys
-            )
-
-            guard let data = decryptedPayload.data(using: .utf8) else {
-                throw AzureError.client("Failed to create utf8 encoded Data from decrypted string.")
-            }
-
-            // 5. Create and return the PushNotificationEvent Model
-            let pushNotificationEvent = try PushNotificationEvent(chatEventType: chatEventType, from: data)
-
-            return pushNotificationEvent
-
-        } catch {
-            throw AzureError.client("Error in decrypting the notification payload: \(error)")
         }
     }
 }
