+++ sdk/communication/AzureCommunicationChat/Source/PushNotification/PushNotificationClient.swift
@@ -153,61 +153,4 @@ internal class PushNotificationClient {
             }
         }
     }
-
-    internal static func decryptPayload(
-        encryptedStr: String,
-        encryptionKeys: [String]
-    ) throws -> String {
-        do {
-            // 1.Decode the Base64 input string into [UInt8].
-            guard let decodedData = Data(base64Encoded: encryptedStr) else {
-                throw AzureError
-                    .client(
-                        "Failed to initialize a data object with the given cryptoKey. Please ensure the encryptedStr is a Base64 encoded string."
-                    )
-            }
-            let encryptedBytes = Array(decodedData)
-
-            // 2.Split [UInt8] into different blocks.
-            let iv: [UInt8] = CryptoUtils.extractInitializationVector(result: encryptedBytes)
-            let cipherText: [UInt8] = CryptoUtils.extractCipherText(result: encryptedBytes)
-            let hmac: [UInt8] = CryptoUtils.extractHmac(result: encryptedBytes)
-            let cipherModeIVCipherText: [UInt8] = CryptoUtils.extractCipherModeIVCipherText(result: encryptedBytes)
-
-            // 3.Loop over the key array and find the key used for encryption
-            for encryptionKey in encryptionKeys {
-                let encryptionKeys = splitEncryptionKey(encryptionKey: encryptionKey)
-                let aesKey = encryptionKeys[0]
-                let authKey = encryptionKeys[1]
-
-                // Each auth key can be computed into a unique HMAC signature. If the computed signature matches the included signature, "verifyHMACResult" will be true.
-                let verifyHMACResult = try verifyEncryptedPayload(
-                    cipherModeIVCipherText: cipherModeIVCipherText,
-                    authKey: authKey,
-                    actualHmac: hmac
-                )
-
-                // If the "verifyHMACResult" is true, it indicates that the current keyPair is the one used by PNH for encryption.
-                if verifyHMACResult {
-                    // Here we use aesKey to decrypt the payload.
-                    let decryptedString = try decryptPushNotificationPayload(
-                        cipherText: cipherText,
-                        iv: iv,
-                        cryptoKey: aesKey
-                    )
-
-                    return decryptedString
-                }
-            }
-
-            // 4. Failed to decrypt the push notification if thre is no keyPair used for encryption.
-            throw AzureError
-                .client(
-                    "Failed to decrypt the push notification."
-                )
-
-        } catch {
-            throw AzureError.client("Error in decrypting the notification payload: \(error)")
-        }
-    }
 }
