+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_StartPushNotifications_ReturnsSuccess.json
@@ -2,12 +2,20 @@
   "name" : "test_StartPushNotifications_ReturnsSuccess",
   "interactions" : [
     {
-      "request" : {
+      "response" : {
         "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "headers" : {
-          "Content-Type" : "application\/json",
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)"
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "x-cache" : "CONFIG_NOCACHE",
+          "ms-cv" : "htzl4Phph063QgowPASA2w.0",
+          "x-msedge-ref" : "Ref A: F681EF34ACB349F2AAE0F09F81C3A34A Ref B: WSTEDGE0622 Ref C: 2022-06-22T23:52:52Z",
+          "x-content-type-options" : "nosniff",
+          "Content-Length" : "0"
         },
+        "status" : 202,
+        "body" : ""
+      },
+      "request" : {
         "method" : "POST",
         "body" : {
           "clientDescription" : {
@@ -31,22 +39,14 @@
           },
           "registrationId" : "0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
           "nodeId" : ""
-        }
-      },
-      "recorded_at" : 1655159407.155288,
-      "response" : {
-        "headers" : {
-          "Content-Length" : "0",
-          "x-msedge-ref" : "Ref A: FF3D967FE9F043F8B2C7540963ED773C Ref B: WSTEDGE1103 Ref C: 2022-06-13T22:30:06Z",
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
-          "x-content-type-options" : "nosniff",
-          "x-cache" : "CONFIG_NOCACHE",
-          "ms-cv" : "LC9ysjq7KEK4SVEvJLx+wQ.0"
         },
-        "status" : 202,
-        "body" : "",
+        "headers" : {
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
+          "Content-Type" : "application\/json"
+        },
         "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations"
-      }
+      },
+      "recorded_at" : 1655941972.709296
     }
   ]
 }
