+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_StopPushNotifications_ReturnsSuccess.json
@@ -1,11 +1,9 @@
 {
+  "name" : "test_StopPushNotifications_ReturnsSuccess",
   "interactions" : [
     {
+      "recorded_at" : 1655941973.0355759,
       "request" : {
-        "headers" : {
-          "Content-Type" : "application\/json",
-          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)"
-        },
         "body" : {
           "clientDescription" : {
             "templateKey" : "AcsIos.AcsNotify_Chat_3.0",
@@ -29,48 +27,50 @@
           "registrationId" : "0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
           "nodeId" : ""
         },
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
-        "method" : "POST"
+        "headers" : {
+          "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)",
+          "Content-Type" : "application\/json"
+        },
+        "method" : "POST",
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations"
       },
-      "recorded_at" : 1655159407.6182971,
       "response" : {
+        "status" : 202,
         "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "headers" : {
-          "x-content-type-options" : "nosniff",
+          "x-cache" : "CONFIG_NOCACHE",
           "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
-          "ms-cv" : "dQwe4fGXWUuna3IbcA2CJQ.0",
-          "x-msedge-ref" : "Ref A: 0FBFC6744FCE4D31ABDADE5EF21D32C8 Ref B: WSTEDGE1103 Ref C: 2022-06-13T22:30:07Z",
+          "ms-cv" : "DS4bQdjLm0WE2xsLuq2cuA.0",
           "Content-Length" : "0",
-          "x-cache" : "CONFIG_NOCACHE"
+          "x-msedge-ref" : "Ref A: 4C245B6F6CF2404CB03BF33A9072A8C2 Ref B: WSTEDGE0622 Ref C: 2022-06-22T23:52:52Z",
+          "x-content-type-options" : "nosniff"
         },
-        "status" : 202,
         "body" : ""
       }
     },
     {
-      "recorded_at" : 1655159407.8874741,
-      "response" : {
-        "headers" : {
-          "x-content-type-options" : "nosniff",
-          "ms-cv" : "4qabVt5\/GUqBzz\/lTAMjcw.0",
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
-          "x-msedge-ref" : "Ref A: 25F17ABE9DFE471C94C2044DBBACF64F Ref B: WSTEDGE1103 Ref C: 2022-06-13T22:30:07Z",
-          "x-cache" : "CONFIG_NOCACHE",
-          "Content-Length" : "0"
-        },
-        "status" : 202,
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
-        "body" : ""
-      },
       "request" : {
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
         "method" : "DELETE",
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
         "headers" : {
           "Content-Type" : "application\/json",
           "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)"
         }
+      },
+      "recorded_at" : 1655941973.305275,
+      "response" : {
+        "body" : "",
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations\/0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
+        "headers" : {
+          "Content-Length" : "0",
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "x-cache" : "CONFIG_NOCACHE",
+          "x-content-type-options" : "nosniff",
+          "ms-cv" : "GwaTKDiFTESoj9kapjzroQ.0",
+          "x-msedge-ref" : "Ref A: D2F999961FC04540A9FB4E677228826B Ref B: WSTEDGE0622 Ref C: 2022-06-22T23:52:53Z"
+        },
+        "status" : 202
       }
     }
-  ],
-  "name" : "test_StopPushNotifications_ReturnsSuccess"
+  ]
 }
