+++ sdk/communication/AzureCommunicationChat/Tests/Recordings/test_setRegistration_ReturnsSuccess.json
@@ -1,13 +1,25 @@
 {
+  "name" : "test_setRegistration_ReturnsSuccess",
   "interactions" : [
     {
+      "response" : {
+        "status" : 202,
+        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
+        "headers" : {
+          "x-cache" : "CONFIG_NOCACHE",
+          "ms-cv" : "ftP08TXzKkartBGTWvMdOw.0",
+          "x-msedge-ref" : "Ref A: D7BF7EBAC36C455F9BBE78DDBEDBFCB3 Ref B: WSTEDGE0605 Ref C: 2022-06-22T23:49:49Z",
+          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
+          "Content-Length" : "0",
+          "x-content-type-options" : "nosniff"
+        },
+        "body" : ""
+      },
       "request" : {
-        "method" : "POST",
         "headers" : {
           "Content-Type" : "application\/json",
           "User-Agent" : "azsdk-ios-AzureCommunicationChatTests\/1.0 (MacBookPro16,1 - 21.5.0; xctest:13.3 -> iOS 9.0; en_US)"
         },
-        "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations",
         "body" : {
           "clientDescription" : {
             "templateKey" : "AcsIos.AcsNotify_Chat_3.0",
@@ -30,23 +42,11 @@
           },
           "registrationId" : "0E0F0BE0-0000-00C0-B000-A00A00E00BD0",
           "nodeId" : ""
-        }
-      },
-      "recorded_at" : 1655161286.3781052,
-      "response" : {
-        "body" : "",
-        "headers" : {
-          "Strict-Transport-Security" : "max-age=31536000; includeSubDomains",
-          "ms-cv" : "72g67SuEuk63Oz4OIX6ktg.0",
-          "x-msedge-ref" : "Ref A: D6B39765D01A4379B401F0E7ECDE6500 Ref B: CO1EDGE1914 Ref C: 2022-06-13T23:01:26Z",
-          "x-cache" : "CONFIG_NOCACHE",
-          "Content-Length" : "0",
-          "x-content-type-options" : "nosniff"
         },
-        "status" : 202,
+        "method" : "POST",
         "url" : "https:\/\/edge.skype.com\/registrar\/prod\/v2\/registrations"
-      }
+      },
+      "recorded_at" : 1655941789.854255
     }
-  ],
-  "name" : "test_setRegistration_ReturnsSuccess"
+  ]
 }
